Refactor constraint traits
