+++ jolt-core/src/r1cs/mod.rs
@@ -1,7 +1,7 @@
 pub mod inputs;
 
 pub mod builder;
-pub mod jolt_constraints;
+pub mod constraints;
 pub mod key;
 pub mod ops;
 pub mod spartan;
