+++ jolt-core/src/r1cs/ops.rs
@@ -11,34 +11,7 @@ use rayon::prelude::*;
 use std::fmt::Debug;
 use std::hash::Hash;
 
-pub trait ConstraintInput:
-    Clone + Copy + Debug + PartialEq + Eq + PartialOrd + Ord + Hash + Sync + Send + 'static
-{
-    // TODO(moodlezoup): Move flattened version to r1cs preprocesing
-    fn flatten<const C: usize>() -> Vec<Self>;
-    fn num_inputs<const C: usize>() -> usize {
-        Self::flatten::<C>().len()
-    }
-    fn from_index<const C: usize>(index: usize) -> Self {
-        Self::flatten::<C>()[index]
-    }
-    fn to_index<const C: usize>(&self) -> usize {
-        match Self::flatten::<C>().iter().position(|x| x == self) {
-            Some(index) => index,
-            None => panic!("Invalid JoltIn variant {:?}", self),
-        }
-    }
-
-    fn get_poly_ref<'a, F: JoltField, PCS: CommitmentScheme<Field = F>>(
-        &self,
-        jolt_polynomials: &'a JoltPolynomials<F, PCS>,
-    ) -> &'a DensePolynomial<F>;
-
-    fn get_poly_ref_mut<F: JoltField, PCS: CommitmentScheme<Field = F>>(
-        &self,
-        jolt_polynomials: &mut JoltPolynomials<F, PCS>,
-    ) -> &mut DensePolynomial<F>;
-}
+use super::inputs::ConstraintInput;
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub enum Variable {
